Clean it up with proper header etc
